did not compile with flags

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: oozsertt <oozsertt@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/17 15:23:18 by oozsertt          #+#    #+#             */
-/*   Updated: 2020/10/12 19:57:38 by oozsertt         ###   ########.fr       */
+/*   Updated: 2020/10/15 00:41:47 by oozsertt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int		i;
+	size_t	i;
 	char	*tempforfill;
 
 	i = 0;
